+++ metadata-mis/src/main/java/com/mingdong/mis/domain/mapper/StatsClientMapper.java
@@ -7,14 +7,14 @@ import java.util.Date;
 
 public interface StatsClientMapper
 {
-    Long getAllClientCount();
+    Integer getAllClientCount();
 
-    Long getClientCountByDate(@Param("start") Date start, @Param("end") Date end);
+    Integer getClientCountByDate(@Param("start") Date start, @Param("end") Date end);
 
     BigDecimal getClientRechargeByDate(@Param("start") Date start, @Param("end") Date end);
 
     BigDecimal getClientRechargeAll();
 
-    Long countClientRechargeByDate(@Param("start") Date start, @Param("end") Date end);
+    Integer countClientRechargeByDate(@Param("start") Date start, @Param("end") Date end);
 
 }
